backend/requirements.txt
fastapi
uvicorn
anthropic
yt-dlp
python-dotenv
pydantic
httpx
